B3::Procedure::resetReachability() can create dangling references from Upsilons to Phis https://bugs.webkit.org/show_bug.cgi?id=159165
Reviewed by Mark Lam.

You can delete an unreachable block that has a Phi but some prior block may still have an
Upsilon pointing to that Phi. This can happen if the Upsilon precedes a Check that always
exits or it can happen if we remove some successor of a block and this block had an Upsilon
for one of the removed successors. These things are valid IR even if they are not canonical.
Our policy for not-canonical-but-valid IR is that the compiler should still emit valid code
in the end.

The solution is to have Procedure::resetReachability() turn those Upsilons into Nops.

* b3/B3Procedure.cpp:
(JSC::B3::Procedure::resetReachability): Fix the bug.
* b3/B3Validate.h:
* b3/testb3.cpp:
(JSC::B3::testResetReachabilityDanglingReference): Add a test. This always crashes prior to this change.
* dfg/DFGGraph.cpp:
(JSC::DFG::Graph::killUnreachableBlocks): Add a FIXME about a possible similar bug.



Canonical link: https://commits.webkit.org/177263@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@202512 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/JavaScriptCore/b3/B3Procedure.cpp
@@ -36,6 +36,7 @@
 #include "B3DataSection.h"
 #include "B3Dominators.h"
 #include "B3OpaqueByproducts.h"
+#include "B3PhiChildren.h"
 #include "B3StackSlot.h"
 #include "B3ValueInlines.h"
 #include "B3Variable.h"
@@ -172,13 +173,31 @@ void Procedure::resetReachability()
         }
     }
     
+    IndexSet<Value> valuesToDelete;
+    
     B3::resetReachability(
         m_blocks,
         [&] (BasicBlock* deleted) {
             // Gotta delete the values in this block.
             for (Value* value : *deleted)
-                deleteValue(value);
+                valuesToDelete.add(value);
         });
+    
+    if (valuesToDelete.isEmpty())
+        return;
+    
+    for (BasicBlock* block : *this) {
+        for (Value* value : *block) {
+            ASSERT(!valuesToDelete.contains(value)); // The block should have been deleted already.
+            if (UpsilonValue* upsilon = value->as<UpsilonValue>()) {
+                if (valuesToDelete.contains(upsilon->phi()))
+                    upsilon->replaceWithNop();
+            }
+        }
+    }
+    
+    for (Value* value : valuesToDelete.values(values()))
+        deleteValue(value);
 }
 
 void Procedure::invalidateCFG()
